requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
requests==2.25.1
python-dotenv==0.19.2
slack_sdk==3.8.0


